agents/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# OAuth and JWT
python-jose[cryptography]>=3.3.0
httpx>=0.25.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.1
langchain-mcp-adapters>=0.1.0

# MCP (Model Context Protocol)
mcp>=0.1.0

# Azure OpenAI
openai>=1.0.0
langchain-openai>=0.0.1

# Environment and utilities
python-dotenv>=1.0.0
rich>=13.0.0

# Additional dependencies
requests>=2.31.0
